app/requirements.txt
paho-mqtt==2.1.0
emoji==2.14.1
wakeonlan==3.1.0
paramiko==3.5.1
python-dotenv==0.19.2
